Crate derive_setters

This crate provides a macro for automatically generating setter methods for a struct’s fields. This can be used to add setters to a plain data struct, or to help in implementing builders.

For detailed usage, see #[derive(Setters)]

Example

use derive_setters::*;

#[derive(Default, Setters, Debug, PartialEq, Eq)]
struct BasicStruct {
    #[setters(rename = "test")]
    a: u32,
    b: u32,
    c: u32,
}

assert_eq!(
    BasicStruct::default().test(30).b(10).c(20),
    BasicStruct { a: 30, b: 10, c: 20 },
);

Derive Macros

Setters

This macro can be used on a struct to generate setter methods for its fields.