derive_setters

Derive Macro Setters

#[derive(Setters)]
{
    // Attributes available to this derive:
    #[setters]
}

This macro can be used on a struct to generate setter methods for its fields.

By default, it generates a single method with the same name as each field, that takes self by value and returns a new struct with the field modified.

The way these methods are generated can be additionally customized with the #[setters(…)] attribute.

Example

A basic usage example:

#[derive(Default, Setters, Debug, PartialEq, Eq)]
struct BasicStruct {
    a: u32,
    b: u32,
    c: u32,
}

This would generate the following implementations:

impl BasicStruct {
    #[must_use]
    pub fn a(mut self, value: u32) -> Self {
        self.a = value;
        self
    }

    #[must_use]
    pub fn b(mut self, value: u32) -> Self {
        self.b = value;
        self
    }

    #[must_use]
    pub fn c(mut self, value: u32) -> Self {
        self.c = value;
        self
    }
}

Struct-level options

The following options control which fields are actually generated:

#[setters(generate = false)] causes setter methods to not be generated by default. This can be overwritten with #[setters(generate)] on the fields you want setters for.

#[setters(generate_private = false)] causes setter methods to not be generated by default for private fields.

#[setters(generate_public = false)] causes setter methods to not be generated by default for public fields.

#[setters(no_std)] causes the generated code to use core instead of std.

#[setters(prefix = "…")] causes the setter method on all fields to be prefixed with the given string.

Field-level options

The following options can be set on a field.

#[setters(generate)] causes a setter method to be generated, overriding struct-level settings.

#[setters(skip)] causes no setter method to be generated.

#[setters(rename = "…")] causes the setter method to have a different name from the field. When used, the prefix option is ignored.

Common options

The following options can be set on either the entire struct, or on an individual field. You can disable the features for a particular field with #[setters(option = false)]

#[setters(into)] causes the setter method to accept any type that can be converted into the field’s type via the Into trait.

#[setters(strip_option)] causes the setter method to accept T instead of Option<T>. If applied to a field that isn’t wrapped in an Option, it does nothing.

#[setters(bool)] causes the setter method to take no arguments, and to always set the value of the field to true.

#[setters(borrow_self)] causes the generated setter method to borrow &mut self instead of taking self, and return &mut Self rather than Self.

Delegation options

The following options allow additional setters to be generated for wrappers around the type this macro is used on.

#[setters(generate_delegates(ty = "OtherTy", field = "some_field"))] causes all setter methods on this struct to be duplicated on the target struct, instead modifying self.some_field instead of self.

#[setters(generate_delegates(ty = "OtherTy", method = "get_field"))] does the same thing as above, except calling get_field with no arguments instead of directly accessing a field.

#[setters(generate_delegates(ty = "OtherTy", field = "some_field", prefix = "set_"))] adds the prefix set_ to all delegated fields.

#[setters(generate_delegates(ty = "OtherTy<T>", generics = "<T>", method = "get_field"))] does the same thing as above, but it supports generic types.